+++ app/src/main/java/com/boostcamp/dreampicker/data/source/remote/firebase/request/InsertUserRequest.java
@@ -1,0 +1,47 @@
+package com.boostcamp.dreampicker.data.source.remote.firebase.request;
+
+public class InsertUserRequest {
+    private String id;
+    private String name;
+    private String email;
+    private String profileImageUrl;
+
+    public InsertUserRequest(String id, String name, String email, String profileImageUrl) {
+        this.id = id;
+        this.name = name;
+        this.email = email;
+        this.profileImageUrl = profileImageUrl;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getProfileImageUrl() {
+        return profileImageUrl;
+    }
+
+    public void setProfileImageUrl(String profileImageUrl) {
+        this.profileImageUrl = profileImageUrl;
+    }
+}
