--- app/src/main/java/com/boostcamp/dreampicker/data/source/remote/firebase/response/ProfileUserDetailResponse.java
@@ -1,82 +1,0 @@
-package com.boostcamp.dreampicker.data.source.remote.firebase.response;
-
-public class ProfileUserDetailResponse {
-    private String id;
-    private String name;
-    private String email;
-    private String profileImageUrl;
-
-    // TODO. 삭제
-    private int profileImageResource;
-
-    private int feedCount;
-    private int followerCount;
-    private int followingCount;
-
-    public ProfileUserDetailResponse() {
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getProfileImageUrl() {
-        return profileImageUrl;
-    }
-
-    public void setProfileImageUrl(String profileImageUrl) {
-        this.profileImageUrl = profileImageUrl;
-    }
-
-    public int getProfileImageResource() {
-        return profileImageResource;
-    }
-
-    public void setProfileImageResource(int profileImageResource) {
-        this.profileImageResource = profileImageResource;
-    }
-
-    public int getFeedCount() {
-        return feedCount;
-    }
-
-    public void setFeedCount(int feedCount) {
-        this.feedCount = feedCount;
-    }
-
-    public int getFollowerCount() {
-        return followerCount;
-    }
-
-    public void setFollowerCount(int followerCount) {
-        this.followerCount = followerCount;
-    }
-
-    public int getFollowingCount() {
-        return followingCount;
-    }
-
-    public void setFollowingCount(int followingCount) {
-        this.followingCount = followingCount;
-    }
-}
